+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - Fixed an error that occurred when creating a TinyMCE field
+- Fixed a bug where, if no TinyMCE fields existed on an element editor page, TinyMCE fields would not load on slideout editors
 
 ## 2.0.0-alpha.2 - 2024-02-07