+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## 2.0.0-beta.1 - 2024-02-20
 
 ### Added
 - Added `spicyweb\tinymce\fields\TinyMCE::getEntryTypes()`
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - Fixed an error that occurred when creating a TinyMCE field
+- Fixed a bug where, if no TinyMCE fields existed on an element editor page, TinyMCE fields would not load on slideout editors
 
 ## 2.0.0-alpha.2 - 2024-02-07